Pass fetchSize to getNextBatch
This fixes a rather intricate bug: The ResultIterator would use the batch size passed on creation, but the fetch code would it compare with the fetch size specified in the message.  This would lead to fetch reporting isLast to early, when the fetch size in the message is different from in the initial execute message.