Rework server loop to be somewhat async
This way seems more logic.  Unparameterized statements are still a bit of a hack, but with the async changes, it should be possible to not send the statement id, but being able to cancel calls based on the message id.

While working on the loop, change to allow the ConnectionRequest only once and only as the first message.