Add unix listener in .polypheny folder for all modes
The way the interface for the listener directly in the .polypheny
folder is created still is sub-optimal, but it works for now.

To create a UNIX socket, we first have to delete the previous sockets
file.  Unfortunately, this works even when the previous socket is
still active being listened at.  To prevent deleting and rebinding
while another instance of Polypheny is using the path we do the
following: Polypheny creates a file (the socket file name with ".lock"
appended) and locks it exclusively.  This exclusive lock gets released
when the interface is destroyed or on process exit.  That way no other
Polypheny can bind to that path, while another instance is listening.