Set callSuper to true for LogicalTable
With callSuper = false, only the primary key field is used to determine
equality.  This is a problem for LogicalTables that map to sources (not
stores), because for these the primary key field is null.  This in turn
means that all LogicalTables for sources are equal.  This is a problem
for the locking code, because then all source tables are locked as one.

With this change the source tables are not equal to each other and can
be locked separately.